+++ israeli-broadcast-schedule-card/israeli-broadcast-schedule-card.js
@@ -1,100 +1,96 @@
-class israeliBroadcastScheduleCard extends HTMLElement {
-    constructor() {
-        super();
-        this.attachShadow({ mode: 'open' });
-    }
+class IsraeliBroadcastScheduleCard extends HTMLElement {
+  set hass(hass) {
+    if (!this.content) {
+      this.card = document.createElement('ha-card');
+      this.card.dir = 'rtl';
 
-    set hass(hass) {
-
-    }
-
-
-    setConfig(config) {
-        if (!config.channel) {
-            throw new Error('Please define a channel');
-        }
-
-        const root = this.shadowRoot;
-        if (root.lastChild) root.removeChild(root.lastChild);
-        const card = document.createElement('ha-card');
-        const style = document.createElement('style');
-        const content = document.createElement('div');
-        const title = document.createElement("div");
-        title.className = "header";
-        title.style = "font-family: var(--paper-font-headline_-_font-family); -webkit-font-smoothing: var(--paper-font-headline_-_-webkit-font-smoothing); font-size: var(--paper-font-headline_-_font-size); font-weight: var(--paper-font-headline_-_font-weight); letter-spacing: var(--paper-font-headline_-_letter-spacing); line-height: var(--paper-font-headline_-_line-height); line-height: 40px; color: var(--primary-text-color); padding: 4px 0 12px; display: flex; justify-content: space-between;";
-        title.innerHTML = '<div class="name">' + config.title + '</div>';
-        style.textContent = `
-          ha-card {
-            padding: 16px;
-            text-align: right;
-            direction: rtl;
+      const style = document.createElement('style');
+      style.textContent = `
+          .schedule-item {
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+            gap: 0.5rem;
+            padding: 0.5rem 0;
           }
-          ul {
-            list-style-type: none;
-            margin: 0;
-            padding: 0;
-            overflow: hidden;
-           }        
-           li a {
-            display: block;
-            color: var(--primary-text-color);
-            text-align: right;
-            padding: 2px 0;
-            text-decoration: none;
-           }
-           .head {
+          .schedule-item:first-child {
             font-weight: bold;
-           }
+          }
+          .schedule-item__time {
+            color: var(--secondary-text-color);
+          }
         `;
-        content.innerHTML = "בטעינה...";
 
-        card.appendChild(style);
-        if (config.title) {
-            card.appendChild(title);
-        }
-        card.appendChild(content);
-        root.appendChild(card);
+      // Keep default ha-card content style
+      this.card.innerHTML = `<div class="card-content"></div>`;
+      this.content = this.card.querySelector('.card-content');
 
-        this._httpGet("https://cors-anywhere.herokuapp.com/" + this._URL(config.channel),
-            function () {
-                if (this.readyState === 4 && this.status === 200) {
-                    let result = document.createElement('div');
-                    result.innerHTML = this.responseText;
+      this.content.innerHTML = 'בטעינה...';
 
-                    if (config.max) {
-                        let elementList = result.querySelectorAll("li");
-                        for (let i = 0; i < elementList.length; i++) {
-                            if (i > config.max) {
-                                result.firstChild.removeChild(elementList[i]);
-                            }
-                        }
-                    }
-
-                    content.innerHTML = result.innerHTML;
-                }
-
-            });
+      this.card.appendChild(style);
+      this.card.appendChild(this.content);
+      this.appendChild(this.card);
     }
 
-    _httpGet(theUrl, onreadystatechange) {
-        let xmlHttp = new XMLHttpRequest();
-        xmlHttp.open("GET", theUrl, true);
-        xmlHttp.send(null);
-        xmlHttp.onreadystatechange = onreadystatechange
+    this.card.header = this.config.title;
+    this.fetchSchedule().then(this.renderSchedule.bind(this));
+  }
+
+  renderSchedule() {
+    // Clear previous schedule
+    while (this.content.firstChild) {
+      this.content.removeChild(this.content.firstChild);
     }
 
+    // Render new schedule
+    for (const item of this.schedule.slice(0, this.config.max)) {
+      const container = document.createElement('div');
+      container.className = 'schedule-item';
 
-    _URL(channel) {
-        let now = new Date();
-        let start = new Date();
-        start.setHours(0,0,0);
-        let diff = Math.trunc((now.getTime() - start.getTime()) / 1000 / 60 / 30); // in half hours
-        return `https://www.yes.co.il/content/YesChannelsHandler.ashx?action=GetDailyShowsByDayAndChannelCode&dayValue=0&dayPartByHalfHour=${diff}&channelCode=${channel}`;
-    }
+      const time = document.createElement('span');
+      time.className = 'schedule-item__time';
+      time.innerText = item.time;
+      container.appendChild(time);
 
-    getCardSize() {
-        return 1;
+      const title = document.createElement('span');
+      title.className = 'schedule-item__title';
+      title.innerText = item.title;
+      container.appendChild(title);
+
+      this.content.appendChild(container);
     }
+  }
+
+  async fetchSchedule() {
+    return fetch(
+      `https://israeli-broadcast-schedule.vercel.app/api/${
+        this.config.channel
+      }?time=${this.config.start_time || ''}`
+    )
+      .then((res) => res.json())
+      .then((data) => {
+        this.schedule = data.schedule;
+      })
+      .catch((error) => {
+        console.error('Error while fetching schedule', error);
+      });
+  }
+
+  setConfig(config) {
+    if (!config.channel) {
+      throw new Error('Please define a channel');
+    }
+    if (config.start_time && !config.start_time.includes(':')) {
+      throw new Error('Please define a valid start_time, e.g. 20:00');
+    }
+    this.config = config;
+  }
+
+  getCardSize() {
+    return 1;
+  }
 }
-
-customElements.define('israeli-broadcast-schedule-card', israeliBroadcastScheduleCard);
+customElements.define(
+  'israeli-broadcast-schedule-card',
+  IsraeliBroadcastScheduleCard
+);