+++ israeli-broadcast-schedule-card/changelog.md
@@ -1,14 +1,26 @@
+# Changelog
+
+## 1.0.0
+
+- Improve UI and theme support
+- Add `start_time` option, to set start time of the schedule
+- Update API server
+
 ## 0.0.4
+
 - Update server link
 
 ## 0.0.3
+
 - Fixing broken server link
 
 ## 0.0.2
+
 - Hide programs that have been broadcast
 - Add max shows to display
 - Add loading message
-- Add ``max`` option, to set maximum items to display
+- Add `max` option, to set maximum items to display
 
 ## 0.0.1
+
 Initial release